final Client Management code

+++ backend/freelancePM/src/main/java/com/freelance/freelancepm/service/ClientService.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 
-// Single Responsibility: Handles business logic for Clients
+// Single Responsibility - Handles business logic for Clients
 @Service
 public class ClientService {
 
@@ -30,12 +30,12 @@ public class ClientService {
     }
 
     // Retrieve client by ID
-    public Optional<Client> getClientById(Long id) {
+    public Optional<Client> getClientById(Integer id) {
         return clientRepository.findById(id);
     }
 
     // Delete client by ID
-    public void deleteClientById(Long id) {
+    public void deleteClientById(Integer id) {
         clientRepository.deleteById(id);
     }
 }
